Revert "Enchantment data should be read as shorts."
There is a problem serialising/deserialising shorts. This will be reverted until a proper fix is in place.

This reverts commit 268e068206c49f1bedbba4d83f4a0d560c1fc1a5 and fixes #685.

+++ src/main/java/io/github/nucleuspowered/nucleus/configurate/typeserialisers/ItemStackSnapshotSerialiser.java
@@ -13,8 +13,8 @@ import ninja.leaping.configurate.objectmapping.ObjectMappingException;
 import ninja.leaping.configurate.objectmapping.serialize.TypeSerializer;
 import org.spongepowered.api.Sponge;
 import org.spongepowered.api.data.DataContainer;
-import org.spongepowered.api.data.key.Keys;
 import org.spongepowered.api.data.persistence.DataTranslators;
+import org.spongepowered.api.data.key.Keys;
 import org.spongepowered.api.item.inventory.ItemStack;
 import org.spongepowered.api.item.inventory.ItemStackSnapshot;
 
@@ -68,8 +68,8 @@ public class ItemStackSnapshotSerialiser implements TypeSerializer<ItemStackSnap
             } else {
                 enchantments.forEach(x -> {
                     try {
-                        short id = Short.parseShort(x.getNode("id").getString());
-                        short lvl = Short.parseShort(x.getNode("lvl").getString());
+                        int id = Integer.parseInt(x.getNode("id").getString());
+                        int lvl = Integer.parseInt(x.getNode("lvl").getString());
 
                         x.getNode("id").setValue(id);
                         x.getNode("lvl").setValue(lvl);
